Step7 最後の砦途中

+++ src/simwinter/MarketPrice.java
@@ -2,9 +2,9 @@ package simwinter;
 
 public class MarketPrice {
     private String ticker;
-    private long marketPrice;
+    private double marketPrice;
 
-    public MarketPrice(String ticker, long marketPrice) {
+    public MarketPrice(String ticker, double marketPrice) {
         this.ticker = ticker;
         this.marketPrice = marketPrice;
     }
@@ -13,7 +13,7 @@ public class MarketPrice {
         return ticker;
     }
 
-    public long getMarketPrice() {
+    public double getMarketPrice() {
         return marketPrice;
     }
 }
